将menuconfig.py重命名为menukconfig.py以防重名，解决了ci环境下无法找到kconfiglib
