+++ tools/building.py
@@ -128,14 +128,6 @@ def PrepareBuilding(env, root_directory, has_libcpu=False, remove_components = [
     global Env
     global Rtt_Root
 
-    os.environ["RTT_DIR"] = root_directory
-    os.environ["BSP_DIR"] = os.getcwd()
-    if not "PKGS_DIR" in os.environ:
-        if "ENV_ROOT" in os.environ:
-            os.environ["PKGS_DIR"] = os.path.join(os.environ["ENV_ROOT"], "packages")
-        else:
-            os.environ["PKGS_DIR"] = os.path.join(os.getcwd(), "packages")
-
     AddOptions()
 
     Env = env
@@ -147,10 +139,18 @@ def PrepareBuilding(env, root_directory, has_libcpu=False, remove_components = [
 
     # set RTT_ROOT in ENV
     Env['RTT_ROOT'] = Rtt_Root
+    os.environ["RTT_DIR"] = Rtt_Root
     # set BSP_ROOT in ENV
     Env['BSP_ROOT'] = Dir('#').abspath
+    os.environ["BSP_DIR"] = Dir('#').abspath
+    # set PKGS_ROOT in ENV
+    if not "PKGS_DIR" in os.environ:
+        if "ENV_ROOT" in os.environ:
+            os.environ["PKGS_DIR"] = os.path.join(os.environ["ENV_ROOT"], "packages")
+        else:
+            os.environ["PKGS_DIR"] = os.path.join(Dir('#').abspath, "packages")
 
-    sys.path = sys.path + [os.path.join(Rtt_Root, 'tools')]
+    sys.path = sys.path + [os.path.join(Rtt_Root, 'tools'), os.path.join(Rtt_Root, 'tools/kconfiglib')]
 
     # {target_name:(CROSS_TOOL, PLATFORM)}
     tgt_dict = {'mdk':('keil', 'armcc'),
@@ -311,7 +311,7 @@ def PrepareBuilding(env, root_directory, has_libcpu=False, remove_components = [
             print('--global-macros arguments are illegal!')
 
     if GetOption('genconfig'):
-        from genconfig import genconfig
+        from menukconfig import genconfig
         genconfig()
         exit(0)
 
@@ -321,23 +321,23 @@ def PrepareBuilding(env, root_directory, has_libcpu=False, remove_components = [
         exit(0)
 
     if GetOption('menuconfig'):
-        from menuconfig import menuconfig
+        from menukconfig import menuconfig
         menuconfig(Rtt_Root)
         exit(0)
 
     if GetOption('pyconfig-silent'):
-        from menuconfig import guiconfig_silent
+        from menukconfig import guiconfig_silent
         guiconfig_silent(Rtt_Root)
         exit(0)
 
     elif GetOption('pyconfig'):
-        from menuconfig import guiconfig
+        from menukconfig import guiconfig
         guiconfig(Rtt_Root)
         exit(0)
 
     configfn = GetOption('useconfig')
     if configfn:
-        from menuconfig import mk_rtconfig
+        from menukconfig import mk_rtconfig
         mk_rtconfig(configfn)
         exit(0)
